% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{build_data_tbl}
\alias{build_data_tbl}
\title{Prepare a data table for processing,  from the time dataframe, tmax, tmin and rain data to be}
\usage{
build_data_tbl(time_df, tmax_data, tmin_data, rain_data,
  convert_temps_from_K_to_C = FALSE)
}
\arguments{
\item{time_df}{Data frame with time series data, produced from `make_time_dataframe_from_brick`}

\item{tmax_data}{Spatial cell extract for Tmax data, produced from `cell_extract_spatial`}

\item{tmin_data}{Spatial cell extract for Tmin data, produced from `cell_extract_spatial`}

\item{rain_data}{Spatial cell extract for rainfall data, produced from `cell_extract_spatial`}
}
\value{
A data frame based on \code{time_df} with the Tmax, Tmin, and rainfall data
}
\description{
Prepare a data table for processing,  from the time dataframe, tmax, tmin and rain data to be
}
