% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_accumulated_heat}
\alias{calc_accumulated_heat}
\title{Calculate accumulated heat}
\usage{
calc_accumulated_heat(df, GDD_values, from_Austral_month,
  to_Austral_month = 12)
}
\arguments{
\item{df}{Data frame with date information}

\item{GDD_values}{Vector of GDD values from the data frame to process}

\item{from_Austral_month}{Number of the month in the Austral season (1 = Jul ... 12 = Jun) to start calculating from}

\item{to_Austral_month}{Number of the month in the Austral season (1 = Jul ... 12 = Jun) to stop calculating from, defaults to 12}
}
\value{
Vector of calculated results, padded with NA values for dates that are not in the range,
so that length of vector is the same as the length of the vector of GDD values
}
\description{
For vector of GDD values and a data frame containing associated date information,
calculate the accumulated heat (cumulative sum of GDD values) for the
}
\examples{
data <- data.frame(date = ..., Austral_season_month = ..., GDD_values = c(...))
calc_accumulated_heat(data, data$GDD_Values, from_Austral_month = 4, to_Austral_month = 10)
}
