% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_EHF}
\alias{calc_EHF}
\title{Calculate Excess Heat Factor from indicators for Excess Heat and Heat Stress}
\usage{
calc_EHF(EHI_sig, EHI_accl)
}
\arguments{
\item{EHI_sig}{Significant event EHI from \code{calc_EHI_sig}}

\item{EHI_accl}{Acclimatisation EHI from \code{calc_EHI_accl}}
}
\value{
Vector with EHF values for given temperatures, padded with NA values at the start to maintain the same length as the input vector
}
\description{
Calculate Excess Heat Factor from indicators for Excess Heat and Heat Stress
}
\examples{
t <- rnorm(n = 80, mean = 20)
DMT_95_pctile <- 19.2
EHI_sig <- calc_EHI_sig(t, DMT_95_pctile)
EHI_accl <- calc_EHI_accl(t, 30, 3)
calc_EHF(EHI_sig, EHI_accl)
}
