% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_EHI_accl}
\alias{calc_EHI_accl}
\title{Calculate acclimatisation Excess Heat Index}
\usage{
calc_EHI_accl(temps, accl_length = 30, sig_spread = 3)
}
\arguments{
\item{temps}{Vector of daily mean temperatures to process}

\item{accl_length}{Length of acclimatisation period}

\item{sig_spread}{Length of period to average for significance value}
}
\value{
Vector with EHI_accl for given temperatures, padded with NA values at the start to maintain the same length as the input vector
}
\description{
Calculate acclimatisation Excess Heat Index
}
\examples{
t <- rnorm(n = 80, mean = 20)
calc_EHI_accl(t, 30, 3)
}
