% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_EHI_sig}
\alias{calc_EHI_sig}
\title{Calculate significant event Excess Heat Index}
\usage{
calc_EHI_sig(temps, DMT_pc_thresh, sig_spread = 3)
}
\arguments{
\item{temps}{Vector of daily mean temperatures to process}

\item{DMT_pc_thresh}{Threshold of daily temperature for the desired climate reference period (calculated using all days of the year)}

\item{sig_spread}{Length of period to average for significance value}
}
\value{
Vector with EHI_sig for given temperatures, padded with NA values at the start to maintain the same length as the input vector
}
\description{
\code{EHI_sig = (Avg DMT over spread of days) - DMT_pc_thresh} as per Nairn 2013.
}
\examples{
t <- c(20.48, 17.81, 19.06, 16.40, 19.34, 20.79, 19.82, 16.28, 18.25, 19.56)
DMT_95_pctile <- 19.2
calc_EHI_sig(t, DMT_95_pctile)
}
