% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_GDD}
\alias{calc_GDD}
\title{Calculate Growing Degree Days}
\usage{
calc_GDD(T_max, T_min, T_base, T_upper = 1000)
}
\arguments{
\item{T_max}{Daily Maximum Temperature}

\item{T_min}{Daily Minimum Temperature}

\item{T_base}{Base Temperature}
}
\value{
The GDD for the given data per the formula: \code{GDD = (T_max - T_min) / 2 - T_base}
}
\description{
Calculates the GDD from daily temperatures Tmax and Tmin
with respect to a base temperature Tbase.
Averages the max and min temperatures and subtracts the
base temperature. Results with negative values are set
to zero (rather than having a 'negative' GDD).
}
\examples{
calc_GDD(24.12, 15.94, 10) #no need to specify upper limit
calc_GDD(24.12, 15.94, 10, 35) #but can specify an upper limit if desired
}
