% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_GDD_weighted}
\alias{calc_GDD_weighted}
\title{Calculate Weighted Growing Degree Days}
\usage{
calc_GDD_weighted(GDD_base_budb, GDD_base_leaf, GDD_base_10, params)
}
\arguments{
\item{GDD_base_budb}{Vector of GDD values with \code{T_base} as the budbreak base temperature}

\item{GDD_base_leaf}{Vector of GDD values with \code{T_base} as the leaf appearance temperature}

\item{GDD_base_10}{Vector of GDD values with \code{T_base} set to 10 degrees C}
}
\value{
Vector of weighted GDD values
}
\description{
Calculates the cumulative weighted GDD where:
\itemize{
  \item{\code{GDD_base_budb} is used when cumulative GDD (x) meets: \code{0 < x < params[["bracket_budb_leaf"]]}}
  \item{\code{GDD_base_leaf} is used when cumulative GDD (x) meets: \code{params[["bracket_budb_leaf"]] < x < params[["bracket_leaf_pick"]]}}
  \item{\code{GDD_base_10} is used when cumulative GDD (x) meets: \code{x > params[["bracket_leaf_pick"]]}}
}
}
