% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_stats_by_area}
\alias{calc_stats_by_area}
\title{Calculate agroclimatic grape indices and statistics (by area only)}
\usage{
calc_stats_by_area(x, params)
}
\arguments{
\item{x}{Data frame to process}

\item{params}{Named list of common parameters for calculations}
}
\value{
The input data frame, but with extra fields calculated for this grouping
}
\description{
EHF calculations need to persist across years for the same regions, as they are calculated using rolling functions
and if grouped by region and year, then the start of each year will not yet be calculated
}
