% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calc_stats_by_area_year_month}
\alias{calc_stats_by_area_year_month}
\title{Calculate agroclimatic grape indices and statistics (by area, year, month)}
\usage{
calc_stats_by_area_year_month(x, params)
}
\arguments{
\item{x}{Data frame to process}

\item{params}{Named list of common parameters for calculations}
}
\value{
The input data frame, but with extra fields calculated for this grouping
}
\description{
Calculate agroclimatic grape indices and statistics (by area, year, month)
}
