% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calculate_stats_for_cells}
\alias{calculate_stats_for_cells}
\title{Calculate agroclimatic grape indices from the given data table for the shapes and params specified}
\usage{
calculate_stats_for_cells(shapes, data_table, params)
}
\arguments{
\item{data_table}{Data to process}

\item{params}{}
}
\value{
A data frame with the full list of calculated indices and statistics grouped by cells
}
\description{
Calculate agroclimatic grape indices from the given data table for the shapes and params specified
}
