% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calculate_stats_for_generic_area}
\alias{calculate_stats_for_generic_area}
\title{Calculate agroclimatic grape indices from the given data table with the named param list}
\usage{
calculate_stats_for_generic_area(data_table, params)
}
\arguments{
\item{data_table}{Data to process}

\item{params}{}
}
\value{
A data frame with the full list of calculated indices and statistics grouped by a generic area object
}
\description{
Calculate agroclimatic grape indices from the given data table with the named param list
}
