% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{calculate_stats_for_wine_regions}
\alias{calculate_stats_for_wine_regions}
\title{Calculate agroclimatic grape indices from the given data table for the shapes and params specified}
\usage{
calculate_stats_for_wine_regions(shapes, data_table, params)
}
\arguments{
\item{shapes}{List of shapes}

\item{data_table}{Data to process}

\item{params}{}
}
\value{
A data frame with the full list of calculated indices and statistics grouped by wine region
}
\description{
Calculate agroclimatic grape indices from the given data table for the shapes and params specified
}
