% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{cell_extract_spatial}
\alias{cell_extract_spatial}
\title{Spatial cell extract}
\usage{
cell_extract_spatial(raster_filename, spatial_layer, subset_layer = NULL,
  fun_list = list(mean = mean, var = var, n = length))
}
\arguments{
\item{raster_filename}{File to process}

\item{spatial_layer}{Spatial layer to categorise data values by}

\item{subset_layer}{Permits extracting a subset of layers from the raster file}

\item{fun_list}{Functions to apply to each data point in the brick}
}
\value{
A data frame with extracted values from the raster file, grouped by each object in the spatial layer
}
\description{
Identify the cells that are touched by the regions in the spatial layer and extract the data from those cells
}
