% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{cell_extract_spatial_file_list}
\alias{cell_extract_spatial_file_list}
\title{Spatial cell extract for a list of files with the same spatial layer}
\usage{
cell_extract_spatial_file_list(raster_filenames, spatial_layer,
  subset_layer = NULL, fun_list = list(mean = mean, var = var, n = length))
}
\arguments{
\item{raster_filenames}{List of files to process}

\item{spatial_layer}{Spatial layer to categorise data values by}

\item{subset_layer}{Permits extracting a subset of layers from the raster file}

\item{fun_list}{Functions to apply to each data point in the brick}
}
\description{
Apply the \code{cell_extract_spatial} function to a list of files, row binding the result
and
}
