% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{combine_bom_region_collection}
\alias{combine_bom_region_collection}
\title{Combine SF with multiple geometry objects into one}
\usage{
combine_bom_region_collection(sf)
}
\arguments{
\item{sf}{Simple feature with multiple polygons stored in $geometry}
}
\value{
The input simple feature, but with the polygons combined
}
\description{
Combine the geometry objects of simple feature with multiple polygons into one multi-polygon.
Metadata such as Shape_Len and Shape_Area will be aggregated for the final polygon.
}
