% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{expand_grid_to_target_input_list}
\alias{expand_grid_to_target_input_list}
\title{Create a data frame where each row is a model/domain/variable to extract}
\usage{
expand_grid_to_target_input_list(model_list, domain_list, variables_list)
}
\arguments{
\item{model_list}{Vector of models (corresponding to directory structure)}

\item{domain_list}{Vector of domains (corresponding to directory structure)}

\item{variables_list}{A list where the variables of interest a particular domain
are values for a named list with the domain prefixed by "domain_" (see example)}
}
\description{
Use expand.grid on a list of models and domains
}
\examples{
model_list <- c("ERA", "ACCESS1-0")
domain_list <- c("5km", "50km", "tas10km", "sea10km")#'
variables_list <- vector("list")
variables_list[["domain_5km"]] = c("tasmax_day", "tasmin_day")
variables_list[["domain_50km"]] = c("tasmax_day")
variables_list[["domain_sea10km"]] = c("tasmax_day")
variables_list[["domain_tas10km"]] = c("tasmax_day", "tasmin_day")

expand_grid_to_target_list(model_list, domain_list, variables_list)
}
