% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{extract_geom_for_shapes}
\alias{extract_geom_for_shapes}
\title{Extract geometry from vector of shapes}
\usage{
extract_geom_for_shapes(shapes)
}
\arguments{
\item{shapes}{Shapes to process}
}
\value{
Vector of same length as input, with just the st_geometry of each shape for each item
}
\description{
Extract geometry from vector of shapes
}
