% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{find_last_date_under}
\alias{find_last_date_under}
\title{Find last date that value was under a particular threshold}
\usage{
find_last_date_under(df_dates, df_values, threshold)
}
\arguments{
\item{df_dates}{Vector of dates from data frame}

\item{df_values}{Vector of values from data frame}

\item{threshold}{Vector of values from data frame}
}
\value{
The item in the date vector which corresponds to the item in the values vector first being below the threshold value.
If the value vector never drops below the threshold, then NA is returned.
}
\description{
Find last date that value was under a particular threshold
}
\examples{
df <- data.frame( date = seq(as.Date("1961-07-01"), as.Date("1961-07-31"), "days"),
                  vals = seq(from = 31, to = 1, by = 1) )
find_last_date_under(df$date, df$vals, 4)
}
