% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{get_mainland_regions_from_filenames}
\alias{get_mainland_regions_from_filenames}
\title{Get mainland regions from filenames in a directory}
\usage{
get_mainland_regions_from_filenames(dir_with_shape_files)
}
\arguments{
\item{dir_with_shape_files}{Directory containing shape files for mainland regions}
}
\value{
A vector with all mainland region names
}
\description{
Search target directory for files matching "region_<region_name>" and returns unique instances
}
\examples{
my_directory <- '~/Dev/regions'
get_mainland_regions_from_filenames(my_directory)
}
