% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{get_mainland_shapes_from_sf_dir}
\alias{get_mainland_shapes_from_sf_dir}
\title{Load shape files from directory}
\usage{
get_mainland_shapes_from_sf_dir(dir_with_shape_files)
}
\arguments{
\item{dir_with_shape_files}{Directory containing shape files for mainland regions}
}
\value{
List of simple features from the loaded shape files, st_cast to MULTIPLOYGON
}
\description{
Load shape files from directory
}
\examples{
my_directory <- '~/Dev/regions'
get_mainland_regions_from_filenames(my_directory)
}
