% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{get_region_data}
\alias{get_region_data}
\title{Get Region Info from Shapes Data}
\usage{
get_region_data(shapes)
}
\arguments{
\item{shapes}{List of shapes file}
}
\value{
A tibble with a row for each region, with the object_ id and label as columns
}
\description{
Get Region Info from Shapes Data
}
