% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{get_target_cell_indices_for_domain}
\alias{get_target_cell_indices_for_domain}
\title{Get target cell indices for a domain from lookup object}
\usage{
get_target_cell_indices_for_domain(target_domain,
  wine_region_polygon_cells_per_domain_lookup)
}
\arguments{
\item{target_domain}{The domain in the lookup object to get the matching cells for}

\item{wine_region_polygon_cells_per_domain_lookup}{Pre-calculated lookup listing the cells that exist in each region for each domain}
}
\value{
An integer vector of unique cells, not necessarily in sequencial order
}
\description{
Return the cells in the specified domain that cover all of the regions.
Duplicate cells (i.e. cells that exist in multiple regions) will be listed only once.
}
