% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{get_tasmania_shapes_from_bom_shape_file}
\alias{get_tasmania_shapes_from_bom_shape_file}
\title{Extract specified list of regions from BOM shape file}
\usage{
get_tasmania_shapes_from_bom_shape_file(bom_tas_shape_file,
  target_bom_districts_tas)
}
\arguments{
\item{bom_tas_shape_file}{Shape file with BOM districts for Tasmanian regions}

\item{target_bom_districts_tas}{List of regions to extract by label name, in the format list(..., c(prepend = TRUE,  label = "South East")) where prepend = TRUE will prepend 'Tasmania' to the label}
}
\value{
List of simple features of each Tasmanian region in the BOM shape file, st_cast to MULTIPLOYGON
}
\description{
Extract specified list of regions from BOM shape file
}
\examples{
src_bom_district_shapes_tas <- "/mnt/external-data/BoM-Forecast-Regions/Tasmanian_post_gfe_forecast_districts.shp"
target_bom_districts_tas <- list(
  c(prepend = FALSE, label = "King Island"),
  c(prepend = FALSE, label = "Furneaux Islands"),
  c(prepend = TRUE,  label = "North West Coast"),
  c(prepend = TRUE,  label = "North East"),
  c(prepend = TRUE,  label = "Central North"),
  c(prepend = TRUE,  label = "East Coast"),
  c(prepend = TRUE,  label = "Upper Derwent Valley"),
  c(prepend = TRUE,  label = "South East"))
shps_tas <- get_tasmania_shapes_from_bom_shape_file(src_bom_district_shapes_tas, target_bom_districts_tas)
}
