% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{heatwave_detect_ehf}
\alias{heatwave_detect_ehf}
\title{Detect Heatwaves from EHF Method Calculations}
\usage{
heatwave_detect_ehf(EHF_vector, params)
}
\arguments{
\item{EHF_vector}{Vector of logical (TRUE/FALSE) values corresponding to EHF > 0}

\item{min_duration}{Minimum number of days in a row with EHF > 0 that gets classified as a heatwave}
}
\value{
A vector with detected EHF heatwaves (fulfilling the duration criteria) set to TRUE, and isolated days set to FALSE
}
\description{
Detect Heatwaves from EHF Method Calculations
}
\examples{
EHF_gt_zero <- c(FALSE, FALSE, FALSE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE,
                 TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, FALSE, FALSE)
heatwave_detect_ehf(EHF_gt_zero)
}
