% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{heatwave_detect_seasonal}
\alias{heatwave_detect_seasonal}
\title{Detect seasonal heatwaves using R Marine Heatwaves package}
\usage{
heatwave_detect_seasonal(data, params)
}
\arguments{
\item{data}{The data set to search for seasonal heatwaves}

\item{params}{Named vector of common params, which must contain "DMT_threshold_pctile", "climatology_start", "climatology_end", "min_duration"}
}
\value{
The input data, but with the results of the RmarineHeatWaves added
}
\description{
Detect seasonal heatwaves using R Marine Heatwaves package
}
