% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{log_calc}
\alias{log_calc}
\alias{log_calc_stats_by_area}
\alias{log_calc_stats_by_area_year}
\alias{log_calc_stats_by_area_year_month}
\alias{log_calc_stats_by_area_calendar_year}
\title{Helper log function for calculate stats group}
\usage{
log_calc_stats_by_area(x)

log_calc_stats_by_area_year(x)

log_calc_stats_by_area_year_month(x)

log_calc_stats_by_area_calendar_year(x)
}
\arguments{
\item{x}{The data frame being processed, so that it can be returned unaltered}
}
\value{
\code{cat} a status message to console for particular group
}
