% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{make_transformed_spatial_shape}
\alias{make_transformed_spatial_shape}
\title{Make Transformed Spatial Shape}
\usage{
make_transformed_spatial_shape(shape_geometries, brick)
}
\arguments{
\item{shape_geometries}{List of shape geometries to transform}

\item{brick}{Brick of which to use projection}
}
\description{
Transform the given shape geometries onto the projection of brick
}
