% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{produce_climatology_for_market_insights}
\alias{produce_climatology_for_market_insights}
\title{Produce Market Insights Climatology Report}
\usage{
produce_climatology_for_market_insights(market_insights_data, period1, period2)
}
\arguments{
\item{market_insights_data}{as output from produce_data_table_for_market_insights}

\item{period1}{Year range for first period to provide data for, e.g. 1961:1990}

\item{period2}{Year range for second period to provide data for, e.g. 1991:2015}
}
\description{
Produce Market Insights Climatology Report
}
