% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{produce_data_table_for_market_insights}
\alias{produce_data_table_for_market_insights}
\title{Produce Data Summary for Market Insights}
\usage{
produce_data_table_for_market_insights(variables_with_stats, from_year, to_year)
}
\arguments{
\item{to_year}{}
}
\description{
Produce Data Summary for Market Insights
}
