% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{rearrange_stats_output}
\alias{rearrange_stats_output}
\title{Rearrange columns of final calculated stats data frame to group by topical groups}
\usage{
rearrange_stats_output(stats_prep)
}
\arguments{
\item{stats_prep}{A complete set of calculated stats; note: all columns must be present otherwise error will occur}
}
\value{
Data frame with stats ordered by indice groups
}
\description{
Rearrange columns of final calculated stats data frame to group by topical groups
}
