% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{combine_bom_region_collection}
\alias{combine_bom_region_collection}
\title{Combine a list of geometry objects into one multi-polygon}
\usage{
combine_bom_region_collection(sf_collection)
}
\arguments{
\item{sf_collection}{}
}
\value{

}
\description{
Combine a list of geometry objects into one multi-polygon
}
