% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{downy_mildew_pressure_index}
\alias{downy_mildew_pressure_index}
\title{Calculate Downy Mildew Pressure Index}
\usage{
downy_mildew_pressure_index(rain, T_max, T_min, T_threshold)
}
\arguments{
\item{rain}{Daily rainfall}

\item{T_max}{Maximum daily temperature}

\item{T_min}{Minimum daily temperature}

\item{T_threshold}{Tempature above which must be sustained during the day for 10 hours}
}
\value{
A logical vector specifying if the index was calculated to be TRUE of FALSE for the given data points
}
\description{
Approximates temperature curve to be from `T_min` at start of day, rising linearly to linear rising to `T_max` at middle of day,
and then falling linearly to `T_min` againt at the end of the day
}
