% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACGI_functions.R
\name{extract_geom_for_shapes}
\alias{extract_geom_for_shapes}
\title{Extract geometry from list of shapes}
\usage{
extract_geom_for_shapes(shapes_list)
}
\arguments{
\item{shapes_list}{}
}
\description{
Extract geometry from list of shapes
}
